# optimize reactor on_each instrumentation