--- a/spring-cloud-sleuth-autoconfigure/src/main/java/org/springframework/cloud/sleuth/autoconfig/instrument/reactor/SleuthReactorProperties.java
+++ b/spring-cloud-sleuth-autoconfigure/src/main/java/org/springframework/cloud/sleuth/autoconfig/instrument/reactor/SleuthReactorProperties.java
@@ -89,9 +89,7 @@ public class SleuthReactorProperties {
 		/**
 		 * Decorates on each operator, will be less performing, but logging will always
 		 * contain the tracing entries in each operator.
-		 * @deprecated to be removed in Sleuth 4.0.0
 		 */
-		@Deprecated
 		DECORATE_ON_EACH,
 
 		/**