--- a/spring-cloud-sleuth-instrumentation/src/main/java/org/springframework/cloud/sleuth/instrument/reactor/ReactorHooksHelper.java
+++ b/spring-cloud-sleuth-instrumentation/src/main/java/org/springframework/cloud/sleuth/instrument/reactor/ReactorHooksHelper.java
@@ -143,7 +143,7 @@ final class ReactorHooksHelper {
 		}
 	}
 
-	private static boolean isTraceContextPropagator(Publisher<?> current) {
+	static boolean isTraceContextPropagator(Publisher<?> current) {
 		return current instanceof TraceContextPropagator;
 	}