@marcingrzejszczak 2021-02-01
Hi @robotmrv thank you for the PR. Since I'm no reactor expert, can you please shortly explain to me the logic behind those changes and why the performance is improved? cc @simonbasle @bsideup