@bsideup 2021-02-01
@marcingrzejszczak @robotmrv I see a few changes in this PR and would highly recommend splitting. Otherwise, it is almost impossible to follow the change.

That said, `RUN_STYLE` was indeed added for Sleuth-like use cases and should give a nice performance boost 👍 

Also, adding @OlegDokuka for feedback.