@OlegDokuka 2021-02-09
@robotmrv @marcingrzejszczak please do not rely on my feedback since I'm out of the context on what is going on there. I looked at the Lift part and it looks good to me, but apart from that - I don't have any opinion