@marcingrzejszczak 2021-02-09
@bsideup so how do you suggest splitting this?