@tony-clarke-amdocs 2021-02-19
Do we know what the difference in performance will be now, between decorate-on-each, decorate-on-last and manual?