@tony-clarke-amdocs 2021-02-20
> You have the benchmarks attached above

If I am reading them correctly then onEachComplex (2.352) ran faster than onManualComplex (2.354). I am surprised that onEachComplex was faster than onManualComplex.