@marcingrzejszczak 2021-02-22
That's the problem with benchmarks in general. We're running benchmarks ourselves too so I'll compare the results from running those for a week or so